+++ src/app/community-builds/by-collection/page.tsx
@@ -9,20 +9,31 @@ import { DBBuild } from '@/features/build/types'
 import { signIn, useSession } from 'next-auth/react'
 import Link from 'next/link'
 import { useEffect, useState } from 'react'
-import { getBuilds } from './actions'
-import useBuildActions from '@/features/build/hooks/useBuildActions'
-import { toast } from 'react-toastify'
-import { isErrorResponse } from '@/features/error-handling/lib/isErrorResponse'
-import { dbBuildToBuildState } from '@/features/build/lib/dbBuildToBuildState'
+import { getBuildsByCollection } from '@/features/build/actions/getBuildsByCollection'
+import BuildListFilters from '@/features/build/components/BuildListFilters'
+import CommunityBuildFilters from '@/features/filters/components/CommunityBuildFilters'
+import { CommunityBuildFilterProps } from '@/features/filters/types'
+import { DEFAULT_COMMUNITY_BUILD_FILTERS } from '@/features/filters/constants'
 
-const ITEMS_PER_PAGE = 8
+const ITEMS_PER_PAGE = 24
+
+export type SortFilter = 'date created' | 'upvotes'
+const sortFilterOptions: SortFilter[] = ['date created', 'upvotes']
 
 export default function Page() {
   const { data: sessionData } = useSession()
 
   const [builds, setBuilds] = useState<DBBuild[]>([])
   const [totalBuildCount, setTotalBuildCount] = useState<number>(0)
+  const [sortFilter, setSortFilter] = useState<SortFilter>('upvotes')
+  const [communityBuildFilters, setCommunityBuildFilters] =
+    useState<CommunityBuildFilterProps>(DEFAULT_COMMUNITY_BUILD_FILTERS)
+
+  function handleChangeFilters(filters: CommunityBuildFilterProps) {
+    setCommunityBuildFilters(filters)
+  }
   const [isLoading, setIsLoading] = useState(true)
+
   const { itemTrackerStorage } = useLocalStorage()
   const { discoveredItemIds } = itemTrackerStorage
 
@@ -40,48 +51,24 @@ export default function Page() {
     itemsPerPage: ITEMS_PER_PAGE,
   })
 
-  const { handleReportBuild } = useBuildActions()
-
   useEffect(() => {
-    async function getBuildsByCollection() {
-      const response = await getBuilds({
+    async function getItemsAsync() {
+      const response = await getBuildsByCollection({
         itemsPerPage: ITEMS_PER_PAGE,
         pageNumber: currentPage,
+        sortFilter,
         discoveredItemIds,
+        communityBuildFilters,
       })
       setBuilds(response.items)
       setTotalBuildCount(response.totalItemCount)
       setIsLoading(false)
     }
-    getBuildsByCollection()
-  }, [currentPage, discoveredItemIds])
+    getItemsAsync()
+  }, [currentPage, communityBuildFilters, sortFilter, discoveredItemIds])
 
-  async function onReportBuild(buildId: string) {
-    const reportedBuild = builds.find((build) => build.id === buildId)
-
-    if (!reportedBuild) {
-      console.error(`Could not find build with id ${buildId}, report not saved`)
-      return
-    }
-    const newReported = !reportedBuild.reported
-    const response = await handleReportBuild(
-      dbBuildToBuildState(reportedBuild),
-      newReported,
-    )
-
-    if (!response || isErrorResponse(response)) {
-      console.error(response?.errors)
-      toast.error(response?.errors?.[0])
-    } else {
-      toast.success(response.message)
-      const newBuilds = builds.map((build) => {
-        if (build.id === buildId) {
-          build.reported = newReported
-        }
-        return build
-      })
-      setBuilds(newBuilds)
-    }
+  function handleSortFilterChange(filter: string) {
+    setSortFilter(filter as SortFilter)
   }
 
   if (!sessionData?.user) {
@@ -109,6 +96,10 @@ export default function Page() {
         subtitle="Browse all community builds that contain only items in your collection, based on the data in the Item Tracker."
       />
 
+      <div className="mb-8 flex w-full max-w-xl items-center justify-center">
+        <CommunityBuildFilters onUpdate={handleChangeFilters} />
+      </div>
+
       <BuildList
         label="Build Results"
         currentPage={currentPage}
@@ -121,13 +112,19 @@ export default function Page() {
         onPreviousPage={handlePreviousPageClick}
         onNextPage={handleNextPageClick}
         onSpecificPage={handleSpecificPageClick}
-        headerActions={undefined}
+        headerActions={
+          <BuildListFilters
+            filter={sortFilter}
+            onFilterChange={handleSortFilterChange}
+            options={sortFilterOptions}
+          />
+        }
       >
         {builds.map((build) => (
           <div key={build.id} className="h-full w-full">
             <BuildCard
               build={build}
-              onReportBuild={onReportBuild}
+              onReportBuild={undefined}
               footerActions={
                 <div className="flex items-center justify-end gap-2 p-2 text-sm">
                   <Link
